requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
aiosqlite==0.19.0
pydantic==2.5.3
pydantic-settings==2.1.0
aiogram==3.3.0
aiohttp==3.9.1
python-dotenv==1.0.1
yookassa==3.0.0
httpx==0.27.0
cryptography==42.0.0
